requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==3.0.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
celery==5.3.6
dill==0.3.7
distro==1.8.0
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
isort==5.12.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
openai==1.3.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.16.1
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
setuptools==68.2.2
sniffio==1.3.0
sqlparse==0.4.4
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.41.3
pika==1.3.2
